fix bug with interceptors and refactor
