+++ src/pages/SignUp.tsx
@@ -1,7 +1,6 @@
 import { Button, Form, Input, message, Modal } from "antd";
 import { ChangeEvent, useState } from "react";
 import { useNavigate } from "react-router";
-import { signUpApi } from "../api";
 import {
 	maxLoginLength,
 	maxPasswordLength,
@@ -15,6 +14,7 @@ import {
 } from "../constants/constants";
 import { SignUpTypes } from "../types";
 import { ApiErrorHandler } from "../utils/ApiErrorHandler";
+import { signUpApi } from "../api/authApi";
 
 export const SignUp = () => {
 	const [form] = Form.useForm();
